% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{prepare}
\alias{prepare}
\title{Prepare Athlete Monitoring Data}
\usage{
prepare(
  data,
  athlete,
  date,
  variable,
  value,
  day_aggregate = function(x) {     sum(x, na.rm = TRUE) },
  NA_session = NA,
  NA_day = NA,
  acute = 7,
  chronic = 28,
  rolling_fill = NA,
  rolling_estimators = function(x) {     c(mean = mean(x, na.rm = TRUE), sd =
    stats::sd(x, na.rm = TRUE), cv = stats::sd(x, na.rm = TRUE)/mean(x, na.rm = TRUE),
    conf = sum(!is.na(x))/length(x)) },
  posthoc_estimators = function(data) {     return(data) },
  group_summary_estimators = function(x) {     c(median = stats::median(x, na.rm =
    TRUE), lower = stats::quantile(x, 0.25, na.rm = TRUE)[[1]], upper =
    stats::quantile(x, 0.75, na.rm = TRUE)[[1]]) }
)
}
\arguments{
\item{data}{Data frame}

\item{athlete}{Name of the column in the \code{data} where the athlete id or name is located}

\item{date}{Name of the column in the \code{data} where the date is located. \code{date} has
to be of \code{Date} class}

\item{variable}{Name of the column in the \code{data} where the variable name is located}

\item{value}{Name of the column in the \code{data} where the value of \code{variable} is located}

\item{day_aggregate}{Function for aggregating multiple day entries. Defaults is \code{sum}}

\item{NA_session}{What value should be imputed for missing values in \code{value}? Default is \code{NA}}

\item{NA_day}{What value should be imputed for missing days? Default is \code{NA}}

\item{acute}{Duration of the acute rolling window. Default is 7}

\item{chronic}{Duration of the chronic rolling window. Default is 7}

\item{rolling_fill}{Value used to fill start of the rolling windows. Default is \code{NA}}

\item{rolling_estimators}{Function providing rolling estimators. See Details}

\item{posthoc_estimators}{Function providing post-hoc estimators. See Details}

\item{group_summary_estimators}{Function providing group summary estimators. See Details}
}
\value{
Object of class \code{athletemonitoring}
}
\description{
Prepare Athlete Monitoring Data
}
\examples{
data("monitoring")

monitoring$Date <- as.Date(monitoring$Date, "\%Y-\%m-\%d")

prepared_data <- prepare(
  data = monitoring,
  athlete = "Full Name",
  date = "Date",
  variable = "Variable",
  value = "Value")
}
