% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-summary.R
\name{summary.athletemonitoring}
\alias{summary.athletemonitoring}
\title{S3 method for summarizing \code{\link{prepare}} results}
\usage{
\method{summary}{athletemonitoring}(object, ...)
}
\arguments{
\item{object}{Object of class \code{athletemonitoring}}

\item{...}{Extra arguments. Not used}
}
\value{
Tibble with athlete - variable summaries
}
\description{
This function returns long summary of the \code{athletemonitoring} object
}
\examples{
# Load monitoring data set
data("monitoring")

# Filter out only 'Training Load'
monitoring <- monitoring[monitoring$Variable == "Training Load", ]

# Convert column to date format (or use numeric)
monitoring$Date <- as.Date(monitoring$Date, "\%Y-\%m-\%d")

# Run the athlete monitoring data preparation
prepared_data <- prepare(
  data = monitoring,
  athlete = "Full Name",
  date = "Date",
 variable = "Variable",
  value = "Value",
  acute = 7,
  chronic = 42,

  # How should be multiple day entries summarised?
  day_aggregate = function(x) {
    sum(x, na.rm = TRUE)
  },

  # Rolling estimators for Acute and Chronic windows
  rolling_estimators = function(x) {
    c(
      "mean" = mean(x, na.rm = TRUE),
      "sd" = sd(x, na.rm = TRUE),
      "cv" = sd(x, na.rm = TRUE) / mean(x, na.rm = TRUE)
    )
  },

  # Additional estimator post-rolling
  posthoc_estimators = function(data) {
    data$ACD <- data$acute.mean - data$chronic.mean
    data$ACR <- data$acute.mean / data$chronic.mean
    data$ES <- data$ACD / data$chronic.sd

    # Make sure to return the data
    return(data)
  },

  # Group summary estimators
  group_summary_estimators = function(x) {
    c(
      "median" = median(x, na.rm = TRUE),
      "lower" = quantile(x, 0.25, na.rm = TRUE)[[1]],
      "upper" = quantile(x, 0.75, na.rm = TRUE)[[1]]
    )
  }
)

# Get summary
prepared_data
summary(prepared_data)
}
