% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-plotting.R
\name{plot.athletemonitoring}
\alias{plot.athletemonitoring}
\title{S3 method for plotting \code{\link{prepare}} results}
\usage{
\method{plot}{athletemonitoring}(x, type = "line", ...)
}
\arguments{
\item{x}{Object of class \code{athletemonitoring}}

\item{type}{Type of the graph. Use "Bar" or "Line" (default). See Details}

\item{...}{Extra arguments. See Details}
}
\description{
This function plots the selected estimator
}
\examples{
    1+1
}
